moving geth chain creation code to GethNode.ts
